Update on EarlyBuildTest

+++ Assets/Scripts/CurrencyFormatter.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public static class CurrencyFormatter
 {
-    public static string ToRupiah(int amount)
+    public static string ToRupiah(float amount)
     {
         return "Rp " + string.Format("{0:N0}", amount).Replace(",", ".");
     }
